CHORE(FTDA): Reorganise file folder structure
